requirements.txt
Flask
python-dotenv
opencv-python
numpy
Pillow
requests
deepface
scipy
pymongo
mtcnn